Transaction Inputs and Outputs
Transactions now hold a list of inputs (accounts to draw from and how much to draw + verifying signature) and outputs (accounts to send to and how much to send).
When a transaction is added to the tangle, some verification is performed to ensure that the inputs completely cover the outputs, and every input account agreed to the transaction.