Optimization: Lock Step
The child vector of every transaction node is now protected by a monitor which ensures that no writes are allowed while data is being red.
All of the functions in the tangle which may be called concurrently now use iterative for loops instead of range based for loops (we were running into issues with other threads invalidating iterators).