Improved Confidence Walk Set Generation
The walk set is now generated starting at the children and then walking forward. It uses a std::set to ensure that it remains a set. Once the set is built, this node and its children are removed (since they are guaranteed to result in a tip that is approved by the node of interest).